fix: Corrected implementation of _checkToStep
Deleted support for union, intersection and difference. These are operations usdo to create a set of assets, meaning that they could never appear in a reaches, unless in an expr. Therefore, this check_to_step should only take care of the case collect, where it will return the asset pointed by the lhs (where the set operation can be) and checks if the rhs is indeed a step for the specified asset.

Deleted support for subType. This is used to denote a subclass, so in the context of an attack step can only be used inside a collect, where the lhs is the subType and the rhs is the associated step.

Added comments.