fix: Corrected checkStep and subfunctions.
The previous implementation, when given a step, would verify if this step was defined by another asset. However, according to the Java check, we should verify the step hierarchy, i.e. if a step is properly extended from the assets parents. For this, when analysing a single step we only check if it is already defined in the asset. In the post analysis, we verify if the steps are properly inherited.

Added comments.