feat: Added tests for step.
In the current implementation of the analyzer, the verifications of the steps are divided into two functions. The first one is done when processing a step, to verify if it has already been defined for this asset. The second one is done after processing all the language, and verifies hierarchic step definitions. For this reason, there are two test files for steps, one for the intra-processing and the other for the post-processing.