feat: Implemented analyse fields.
Previously, if an asset mentioned a field, the analyzer would check if there was any association with this field. This was incorrect, since an asset was able to use an association it did not belong to. Now, an asset inherits its parents associations and can only use these ones or the ones where it belongs to.

Also changed the post analysis order, as using True or False was previously allowing for it to run twice when it should only happen once: when everything has been parsed.